:sparkles: Make Firefox module find all profiles
