+++ roles/gitlab/tasks/main.yaml
@@ -58,15 +58,23 @@
     name: gitlab-runner
   when: ansible_system == "Win32NT"
 
+- name: Get all groups for Unix
+  ansible.builtin.getent:
+    database: group
+    split: ":"
+  become: true
+  when: ansible_system != "Win32NT"
+
 - name: Configure Gitlab runner user for Unix
   ansible.builtin.user:
     append: true
     name: gitlab-runner
     groups: >-
-      {{ ["sudo"] if ansible_system == "Darwin" else ["docker", "sudo"] }}
+      {{ ['docker', 'sudo', 'wheel'] |
+      intersect(ansible_facts.getent_group.keys()) }}
     shell: /bin/bash
   become: true
-  when: ansible_system in ["FreeBSD", "Linux"]
+  when: ansible_system != "Win32NT"
 
 - name: Enable GitLab runner passwordless sudo for Unix
   ansible.builtin.blockinfile:
